fix: 수익률 계산 오류 수정
- int 자료형 대신 long을 사용해 당첨금이 2,147,483,647이 넘어가는 경우 발생하던 문제 수정
- int 자료형 대신 double 자료형을 사용해 (수익률/입력 금액)의 결과값이 수익률 계산시 소수점 아래의 수까지 반영되도록 수정

+++ src/main/java/lotto/model/Result.java
@@ -35,7 +35,7 @@ public class Result {
         matchResult.replace(LottoPrize.FIVE_MATCH, matchResult.get(LottoPrize.FIVE_MATCH) - ConstantNumber.ONE_CASE);
     }
 
-    public void calculateEarningRate(int earning, int price) {
-        earningRate = (float) (earning / price) * PERCENT;
+    public void calculateEarningRate(long earning, int price) {
+        earningRate = (float) ((double) earning / price) * PERCENT;
     }
 }
